chore(b3fs): rename hash_tree.rs
